requirements.txt
discord.py==1.2.4
dice==2.4.2
python-dotenv==0.10.3
